pyproject.toml
[tool.poetry]
name = "kydo-seed"
version = "0.0.0"
readme = "README.md"
packages = [{include = "kydo_seed", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.12"
google-genai = ">=1.24.0,<2.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# This section is for other tools to read standard metadata
[project]
name = "kydo-seed"
version = "0.0.0"
description = ""
authors = [
    {name = "Your Name", email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.24.0,<2.0.0",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "google-cloud-aiplatform (>=1.101.0,<2.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "shortuuid (>=1.0.13,<2.0.0)",
    "litellm (>=1.74.4,<2.0.0)",
]